feat: add new packet type TRIGGER_IMPORT_BROWSER
Adds a new `UdpPacketType.TRIGGER_IMPORT_BROWSER` for communication from the App to the Server.
This packet type is intended to signal the server to trigger a browser import process.
A new data class `TriggerImportPayload` is introduced to carry the URL of the Ktor server running on the Android device as the payload for this packet type.