Refactor: Update package for ButtonEntity
The `ButtonEntity` class has been moved to the `com.ongxeno.android.starbuttonbox.datasource.room` package.
The `index = true` attribute was removed from the `layout_id` column in `ButtonEntity` as it's already covered by the `indices` definition in the `@Entity` annotation.
An unnecessary import of `ButtonEntity` was removed from `ButtonDao.kt`.